tests: move file fixtures
